requirements.txt
numpy
pandas
scikit-learn
soccerdata
requests
beautifulsoup4
matplotlib
seaborn
PyYAML